this seems to be the happiest state i can get rgbmode into
animation is a bust when i2c is on. bakingpy(nooges, dude behind
keeb.io) is working on it... when he gets a chance. static rainbow is
still fine. had to tweak quantum.c to pick that up on the right side.
so now, rainbows and solid color (with all the colormods) are supported.